Create README.md
Background: 
Our client was an oil and gas consultant who had developed a complex Excel-based model that identified wells that could be purchased based on publicly available data, such as monthly production of oil, gas, and water, vertical and horizontal dimensions, and more. However, the Excel workbook had become so large and prone to crashing due to the sheer size of the data being processed that the client approached us to find an easier and more efficient solution.

Problem: 
The client's field was highly specialized, and it was challenging to understand the logic and steps behind building their model.

Solution: 
After several discussions with the client, we approached the project as an automation challenge. We analyzed the Excel model and examined each tab to determine how they were interdependent. We then identified any steps that could be merged or eliminated to make the model more efficient. Using Python, we coded all the steps in the model to reduce the risk of the workbook crashing.
However, the client was concerned about having to install Python and run the model from a local machine, as it would make it difficult to access from other devices. To solve this issue, we exported the Python code to Google Colab and created a system to save the output in Google Sheets, reducing the need to run the model every time. Additionally, we suggested uploading all the raw data to Google Drive to avoid having to upload it repeatedly.
To help the client build visualizations based on the output generated in Google Sheets, we connected it to Google Data Studio and created an interactive dashboard that was not possible in the earlier Excel model. The dashboard was accessible on any device and could be viewed from anywhere.
Our client was extremely satisfied with the results and even integrated Python into their operations, training their field engineers to use the language. They also recommended us to one of their business associates who required a similar solution.